Re-arrange files and support browser
